feat: make integer and struct emission demand-driven

+++ testdata/projects/integer-constants/expected.ts
@@ -1,4 +1,4 @@
 import type { int64 } from "../../../support/scalars.js";
 export function Small(): int64 {
-    return 42 as int64;
+    return 42;
 }
